requirements.txt
PyYAML
easydict
cython
opencv-python
pandas
tqdm
pycocotools
jpeg4py
tb-nightly
tikzplotlib
thop
colorama
lmdb
scipy
visdom
tensorboardX
setuptools==59.5.0
wandb
timm
pytorch==1.9.0
torchvision==0.10.0
torchaudio==0.9.0
cudatoolkit==10.2
Pillow
einops
